backend/requirements.txt
fastapi
uvicorn
websockets
twilio
pandas
google-genai
python-dotenv
